refactor: move Fetch into common
